Authenticate with Touchstone; use new API
Switch to Touchstone for authentication. The new workflow is:
* Request the URL in the web app that is used for authentication, adding
  `app_url` as a query param to redirect back to the app.
* After authentication the server checks for this param and redirects
  back to this URL. The server adds `token` as a query param; the value
  is a JSON web token (JWT) for authentication subsequent API requests.
* The app stores token using `SecureStore` and fetches the user's
  profile information.

Remove use of sockets to keep the charity in sync.. New API provides an
endpoint to request a sign-in URL that will redirect to the
charity selection form. Similar to signing in, the app adds `app_url` to
this URL opening it in the browser. The web app tracks the value and
displays it on the page to send the user back to the app. The path of
`app_url` specifies where to go in the app.

Add wrapper around API host and switch the value between dev and
production. This and switching to Touchstone removes the need for
`config.js`.

The user status response from the API no longer includes their
send/receive history. It's now requested when rendering the component.
The component was changed to a functional component for simplicity.

Remove `votedCharities` screen and code related to voting for charities;
no longer in use.

Instead of expecting the full user profile from every API request, the
app now queries for the profile after requests that affect it.